Change assertTrue to assertEquals where possible
In a lot of testcases the assertTrue method was used with boolean
expression comparing two values as parameter. This should be avoided as
much as possible, especcially when comparing variables of the type double
or float. When comparing doubles the assertEquals(expected, actual, delta)
method is much better suited because this method takes double inprecision
into account.